% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownToLatex.R
\name{markdownToLatex}
\alias{markdownToLatex}
\title{Convert R markdown to F1000 latex}
\usage{
markdownToLatex(input)
}
\arguments{
\item{input}{path to Rmd file to be converted}
}
\value{
No value is returned, but a tex file is written to disk.  Currently this is 
output to the same location as the input file, and is given an identical name except 
for the file extension.
}
\description{
Turn Rmarkdown version of workflow document into latex file that can be uploaded to F1000 Overleaf
}

