% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToOverleaf.R
\name{uploadToOverleaf}
\alias{uploadToOverleaf}
\title{Upload a LaTeX project to Overleaf}
\usage{
uploadToOverleaf(files = NULL, openInBrowser = FALSE)
}
\arguments{
\item{files}{Character vector of file names to upload.  If the first entry is a zip file this is uploaded directly.
Otherwise the files will be added to a zip archive and then uploaded.}

\item{openInBrowser}{Boolean determining whether to open a browser at the create Overleaf project or not.}
}
\description{
Upload a LaTeX project to Overleaf
}

