% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBiocWorkflow.R
\name{createBiocWorkflow}
\alias{createBiocWorkflow}
\title{Create a New Bioconductor Workflow Package}
\usage{
createBiocWorkflow(
  path,
  description = getOption("devtools.desc"),
  rstudio = TRUE,
  open = rstudio
)
}
\arguments{
\item{path}{location to create new package. The last component of the path will be used as the package name.}

\item{description}{list of description values to override default values or add additional values.}

\item{rstudio}{if \code{TRUE}, creates an RStudio project file.}

\item{open}{if \code{TRUE}, opens the project in a new RStudio session.}
}
\value{
File path to the R Markdown vignette (invisibly).
}
\description{
Uses \code{\link{create_package}} to set up a skeleton for a new Bioconductor workflow package.
}
\examples{
createBiocWorkflow(file.path(tempdir(), "MyWorkflow"), open = FALSE)
}
