% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1000_article.R
\name{f1000_article}
\alias{f1000_article}
\title{F1000 article format}
\usage{
f1000_article(toc = FALSE, number_sections = FALSE, keep_tex = TRUE, ...)
}
\arguments{
\item{toc}{TRUE to include a table of contents in the output}

\item{number_sections}{TRUE to number section headings}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating F1000 software articles.
}
\details{
Creates a latex file that can be uploaded to F1000 Overleaf
}
\examples{

\dontrun{
rmarkdown::draft("MyArticle.Rmd", template="f1000_article", package="BiocWorkflowTools")
}

}

