% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1000_article.R
\name{f1000_article}
\alias{f1000_article}
\title{F1000 article format}
\usage{
f1000_article(toc = FALSE, number_sections = FALSE, fig_width = 5.67,
  fig_height = fig_width, fig_align = "center", keep_tex = TRUE,
  md_extensions = "+link_attributes", pandoc_args = "--wrap=preserve", ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{number_sections}{\code{TRUE} to number section headings}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_align}{Alignment of figures. Possible values are "center" (default) "left" and "right.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating F1000 software articles.
}
\details{
Creates a latex file that can be uploaded to F1000 Overleaf
}
\examples{

\dontrun{
rmarkdown::draft("MyArticle.Rmd", template="f1000_article", package="BiocWorkflowTools")
}

}
