% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1000_article.R
\name{f1000_article}
\alias{f1000_article}
\title{F1000 article format}
\usage{
f1000_article(toc = FALSE, number_sections = FALSE, keep_tex = TRUE,
  md_extensions = "+link_attributes", pandoc_args = "--wrap=preserve",
  extra_dependencies = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{number_sections}{\code{TRUE} to number section headings}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{extra_dependencies}{A LaTeX dependency \code{latex_dependency()}, a
list of LaTeX dependencies, a character vector of LaTeX package names (e.g.
\code{c("framed", "hyperref")}), or a named list of LaTeX package options
with the names being package names (e.g. \code{list(hypreref =
c("unicode=true", "breaklinks=true"), lmodern = NULL)}). It can be used to
add custom LaTeX packages to the .tex header.}

\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating F1000 software articles.
}
\details{
Creates a latex file that can be uploaded to F1000 Overleaf
}
\examples{

\dontrun{
rmarkdown::draft("MyArticle.Rmd", template="f1000_article", package="BiocWorkflowTools")
}

}
