% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadToOverleaf.R
\name{uploadToOverleaf}
\alias{uploadToOverleaf}
\title{Upload a LaTeX project to Overleaf}
\usage{
uploadToOverleaf(files = NULL, openInBrowser = FALSE)
}
\arguments{
\item{files}{Character vector of file names to upload.  If the first entry
is a zip file this is uploaded directly.  Otherwise the files will be added
to a zip archive and then uploaded.}

\item{openInBrowser}{Boolean determining whether to open a browser at the 
created Overleaf project or not. Default value is FALSE.}
}
\value{
No value is returned.  The URL where the uploaded project can be 
accessed is printed to the screen.  If the argument \code{openInBrowser} is 
set to \code{TRUE}, then the default browser will automatically open at the 
Overleaf project page.
}
\description{
Upload a LaTeX project to Overleaf
}
\examples{
example_Rmd <- system.file('examples/f1000_software_example.Rmd', 
                           package = "BiocWorkflowTools")
output_dir <- file.path(tempdir(), 'example')
markdownToLatex(input = example_Rmd, output = output_dir, 
                compress = TRUE)
\dontrun{
## don't run this code chunk in the example as we don't want to spam Overleaf
zip_file <- paste0(output_dir, '.zip')
uploadToOverleaf(files = zip_file, openInBrowser = TRUE)
}                

}

